requirements.txt
Flask==0.10.1
Flask-Assets==0.11
Flask-Script==2.0.5
Flask-WTF==0.12
Jinja2==2.8
MarkupSafe==0.23
WTForms==2.0.2
Werkzeug==0.10.4
beautifulsoup4==4.4.0
cssmin==0.2.0
google-api-python-client==1.4.1
gunicorn==19.3.0
httplib2==0.9.1
icalendar==3.9.0
itsdangerous==0.24
jsmin==2.1.4
oauth2client==1.5.0
py==1.4.30
pyasn1==0.1.8
pyasn1-modules==0.0.7
pytest==2.7.2
python-dateutil==2.4.2
pytz==2015.4
requests==2.7.0
rsa==3.2
simplejson==3.8.0
six==1.9.0
uritemplate==0.6
webassets==0.11.1
